Remove IActivate (#9705)
* git mv

* purge IActivate

* File scoped NS
